# Change decode block sizes