This improves the performance of some of the parquet decode kernels by altering the decode block size:

1) For non-dictionary string columns, the decode kernel block size is reduced 128 -> 64, which improves cuDF benchmark performance by ~10%.  For wide string columns this improvement is up to 25%. 

2) For non-string/non-list chunked reads, the decode kernel block size is increased from 128 -> 256, which improves cuDF benchmark performance by ~10%. This is presumably because the non-list chunked reads have to loop through all of the prior definition levels and count nulls, which is not as bottlenecked on memory usage as the rest of the decode. 

## Checklist
- [ ] I am familiar with the [Contributing Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [ ] New or existing tests cover these changes.
- [ ] The documentation is up to date with these changes.