@pmattione-nvidia 2025-09-15
@mhaseeb123 Relevant benchmarks on 24GB A5000:
BEFORE:
## parquet_read_decode
| data_type |    io_type    | cardinality | run_length | null_probability | data_size | Samples | CPU Time  | Noise | GPU Time  | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|-----------|---------------|-------------|------------|------------------|-----------|---------|-----------|-------|-----------|-------|------------------|-------------------|-------------------|
|    STRING | DEVICE_BUFFER |           0 |          1 |                0 | 536870912 |    509x | 29.413 ms | 0.66% | 29.413 ms | 0.66% |      18252684210 |         1.474 GiB |       482.854 MiB |
|    STRING | DEVICE_BUFFER |        1000 |          1 |                0 | 536870912 |     69x |  7.269 ms | 0.37% |  7.266 ms | 0.37% |      73887151058 |       553.638 MiB |        33.125 MiB |
|    STRING | DEVICE_BUFFER |           0 |         32 |                0 | 536870912 |    234x | 29.691 ms | 0.50% | 29.691 ms | 0.50% |      18081665405 |         1.474 GiB |       482.854 MiB |
|    STRING | DEVICE_BUFFER |        1000 |         32 |                0 | 536870912 |    866x |  7.537 ms | 0.50% |  7.534 ms | 0.50% |      71255449000 |       528.480 MiB |         3.679 MiB |
|    STRING | DEVICE_BUFFER |           0 |          1 |             0.01 | 536870912 |    119x | 30.597 ms | 0.50% | 30.597 ms | 0.50% |      17546598506 |         1.463 GiB |       478.839 MiB |
|    STRING | DEVICE_BUFFER |        1000 |          1 |             0.01 | 536870912 |    720x |  8.216 ms | 0.54% |  8.214 ms | 0.54% |      65361179476 |       571.125 MiB |        33.910 MiB |
|    STRING | DEVICE_BUFFER |           0 |         32 |             0.01 | 536870912 |     72x | 30.770 ms | 0.50% | 30.770 ms | 0.50% |      17447646424 |         1.463 GiB |       478.839 MiB |
|    STRING | DEVICE_BUFFER |        1000 |         32 |             0.01 | 536870912 |    163x |  8.087 ms | 0.50% |  8.085 ms | 0.50% |      66405332246 |       525.479 MiB |         3.696 MiB |

## parquet_read_chunks
|     T     |    io_type    | cardinality | run_length | chunk_read_limit | data_size | Samples |  CPU Time  | Noise |  GPU Time  | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|-----------|---------------|-------------|------------|------------------|-----------|---------|------------|-------|------------|-------|------------------|-------------------|-------------------|
|  INTEGRAL | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      5x | 170.902 ms | 0.16% | 170.921 ms | 0.16% |       3141039267 |       546.959 MiB |       477.515 MiB |
|  INTEGRAL | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      5x | 162.698 ms | 0.22% | 162.716 ms | 0.22% |       3299444498 |       219.219 MiB |       155.184 MiB |
|  INTEGRAL | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |      5x | 144.421 ms | 0.17% | 144.436 ms | 0.17% |       3717008029 |        68.716 MiB |        25.837 MiB |
|  INTEGRAL | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |      5x | 141.600 ms | 0.16% | 141.615 ms | 0.16% |       3791049370 |        64.789 MiB |        13.356 MiB |
|     FLOAT | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      9x |  57.557 ms | 0.14% |  57.561 ms | 0.14% |       9326990709 |       525.910 MiB |       499.881 MiB |
|     FLOAT | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      7x |  78.335 ms | 0.33% |  78.342 ms | 0.33% |       6852937655 |       146.507 MiB |       107.339 MiB |
|     FLOAT | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |      8x |  67.240 ms | 0.06% |  67.245 ms | 0.06% |       7983805098 |        48.752 MiB |        22.186 MiB |
|     FLOAT | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |      8x |  66.141 ms | 0.13% |  66.146 ms | 0.13% |       8116511701 |        44.719 MiB |         9.064 MiB |
|     BOOL8 | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      5x | 807.409 ms | 0.02% | 807.514 ms | 0.02% |        664844412 |       248.081 MiB |        71.665 MiB |
|     BOOL8 | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      5x | 810.237 ms | 0.13% | 810.342 ms | 0.13% |        662523789 |       243.917 MiB |        70.101 MiB |
|     BOOL8 | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |      5x | 795.514 ms | 0.17% | 795.617 ms | 0.17% |        674785649 |       181.249 MiB |        21.573 MiB |
|     BOOL8 | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |      5x | 794.745 ms | 0.03% | 794.848 ms | 0.03% |        675438309 |       181.082 MiB |        21.467 MiB |
|   DECIMAL | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |     13x |  39.925 ms | 0.18% |  39.927 ms | 0.18% |      13446473732 |       701.863 MiB |       393.779 MiB |
|   DECIMAL | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |     11x |  48.518 ms | 0.38% |  48.520 ms | 0.38% |      11064876631 |       114.855 MiB |        70.082 MiB |
|   DECIMAL | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |     12x |  42.808 ms | 0.30% |  42.810 ms | 0.31% |      12540880465 |        48.991 MiB |        15.893 MiB |
|   DECIMAL | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |     13x |  40.161 ms | 0.32% |  40.163 ms | 0.32% |      13367319813 |        35.705 MiB |         5.885 MiB |
| TIMESTAMP | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      8x |  70.000 ms | 0.23% |  70.006 ms | 0.23% |       7668927063 |       704.025 MiB |       453.009 MiB |
| TIMESTAMP | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      8x |  63.003 ms | 0.34% |  63.008 ms | 0.34% |       8520713094 |       129.460 MiB |        90.129 MiB |
| TIMESTAMP | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |      9x |  56.064 ms | 0.22% |  56.068 ms | 0.22% |       9575321877 |        55.376 MiB |        19.251 MiB |
| TIMESTAMP | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |     10x |  52.738 ms | 0.20% |  52.741 ms | 0.20% |      10179334104 |        40.685 MiB |         7.787 MiB |
|  DURATION | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      9x |  62.318 ms | 0.18% |  62.322 ms | 0.18% |       8614400011 |       488.432 MiB |       349.613 MiB |
|  DURATION | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      8x |  62.910 ms | 0.31% |  62.915 ms | 0.31% |       8533275729 |       127.570 MiB |        88.156 MiB |
|  DURATION | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |     10x |  53.577 ms | 0.20% |  53.580 ms | 0.20% |      10019945298 |        42.128 MiB |        16.076 MiB |
|  DURATION | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |     10x |  52.775 ms | 0.26% |  52.778 ms | 0.26% |      10172231393 |        41.225 MiB |         7.565 MiB |
|    STRUCT | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      5x | 171.794 ms | 0.29% | 171.814 ms | 0.29% |       3124719213 |       916.791 MiB |       499.298 MiB |
|    STRUCT | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      7x |  74.586 ms | 0.14% |  74.592 ms | 0.14% |       7197403769 |       199.986 MiB |        74.356 MiB |
|    STRUCT | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |      5x | 172.608 ms | 0.48% | 172.628 ms | 0.48% |       3109994480 |       785.742 MiB |       358.247 MiB |
|    STRUCT | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |      7x |  73.629 ms | 0.17% |  73.635 ms | 0.17% |       7290975107 |        69.781 MiB |         9.099 MiB |

## parquet_read_wide_tables
| data_type | data_size  | num_cols | cardinality | run_length | Samples | CPU Time  | Noise | GPU Time  | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|-----------|------------|----------|-------------|------------|---------|-----------|-------|-----------|-------|------------------|-------------------|-------------------|
|    STRING | 1073741824 |      256 |           0 |          1 |    387x | 38.574 ms | 0.63% | 38.576 ms | 0.63% |      27834464138 |         1.991 GiB |      1010.338 MiB |
|    STRING | 1073741824 |      512 |           0 |          1 |    376x | 39.536 ms | 0.69% | 39.538 ms | 0.69% |      27157039070 |         1.992 GiB |      1010.454 MiB |
|    STRING | 1073741824 |      256 |        1000 |          1 |    528x | 15.118 ms | 0.85% | 15.117 ms | 0.85% |      71029767422 |         1.080 GiB |        70.339 MiB |
|    STRING | 1073741824 |      512 |        1000 |          1 |    544x | 16.416 ms | 0.94% | 16.414 ms | 0.94% |      65416261136 |         1.089 GiB |        75.201 MiB |
|    STRING | 1073741824 |      256 |           0 |         32 |    258x | 39.030 ms | 0.50% | 39.032 ms | 0.50% |      27509381079 |         1.991 GiB |      1010.338 MiB |
|    STRING | 1073741824 |      512 |           0 |         32 |    374x | 39.719 ms | 0.59% | 39.721 ms | 0.59% |      27032289751 |         1.992 GiB |      1010.454 MiB |
|    STRING | 1073741824 |      256 |        1000 |         32 |    544x | 15.471 ms | 0.72% | 15.445 ms | 0.65% |      69518815012 |         1.020 GiB |        10.075 MiB |
|    STRING | 1073741824 |      512 |        1000 |         32 |    152x | 16.242 ms | 0.50% | 16.165 ms | 0.50% |      66425299454 |         1.028 GiB |        14.354 MiB |

## parquet_read_long_strings
|    io_type    | cardinality | data_size | avg_string_length | Samples | CPU Time  | Noise | GPU Time  | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|---------------|-------------|-----------|-------------------|---------|-----------|-------|-----------|-------|------------------|-------------------|-------------------|
| DEVICE_BUFFER |           0 | 536870912 |          2^4 = 16 |     16x | 32.369 ms | 0.43% | 32.262 ms | 0.43% |      16641164920 |         1.664 GiB |       472.933 MiB |
| DEVICE_BUFFER |        1000 | 536870912 |          2^4 = 16 |     57x |  8.938 ms | 0.39% |  8.906 ms | 0.39% |      60279318488 |       781.873 MiB |        33.899 MiB |
| DEVICE_BUFFER |           0 | 536870912 |          2^6 = 64 |     23x | 22.264 ms | 0.20% | 22.192 ms | 0.20% |      24191849450 |         1.559 GiB |       507.947 MiB |
| DEVICE_BUFFER |        1000 | 536870912 |          2^6 = 64 |   1568x |  7.033 ms | 0.53% |  7.010 ms | 0.52% |      76581510745 |       603.609 MiB |        13.661 MiB |
| DEVICE_BUFFER |           0 | 536870912 |         2^8 = 256 |     65x |  7.829 ms | 0.36% |  7.808 ms | 0.36% |      68762350994 |         1.020 GiB |       511.204 MiB |
| DEVICE_BUFFER |        1000 | 536870912 |         2^8 = 256 |    156x | 11.899 ms | 0.50% | 11.870 ms | 0.50% |      45229315134 |       566.420 MiB |        18.160 MiB |
| DEVICE_BUFFER |           0 | 536870912 |       2^10 = 1024 |    576x |  6.910 ms | 0.61% |  6.892 ms | 0.61% |      77894937055 |         1.008 GiB |       511.576 MiB |
| DEVICE_BUFFER |        1000 | 536870912 |       2^10 = 1024 |     87x | 14.182 ms | 0.50% | 14.152 ms | 0.50% |      37935839006 |       584.507 MiB |        62.871 MiB |
| DEVICE_BUFFER |           0 | 536870912 |       2^12 = 4096 |     87x |  5.823 ms | 0.22% |  5.809 ms | 0.22% |      92416102033 |         1.006 GiB |       512.398 MiB |
| DEVICE_BUFFER |        1000 | 536870912 |       2^12 = 4096 |     59x |  8.635 ms | 0.14% |  8.617 ms | 0.14% |      62305024843 |         1.101 GiB |       513.026 MiB |
| DEVICE_BUFFER |           0 | 536870912 |      2^14 = 16384 |     84x |  6.014 ms | 0.43% |  6.001 ms | 0.43% |      89463148872 |         1.007 GiB |       514.864 MiB |
| DEVICE_BUFFER |        1000 | 536870912 |      2^14 = 16384 |     84x |  6.029 ms | 0.29% |  6.016 ms | 0.29% |      89237120389 |         1.008 GiB |       515.432 MiB |
| DEVICE_BUFFER |           0 | 536870912 |      2^16 = 65536 |    213x |  7.311 ms | 0.50% |  7.297 ms | 0.50% |      73571073502 |         1.003 GiB |       518.501 MiB |
| DEVICE_BUFFER |        1000 | 536870912 |      2^16 = 65536 |    134x |  7.318 ms | 0.50% |  7.305 ms | 0.50% |      73495521856 |         1.004 GiB |       519.369 MiB |

AFTER:
## parquet_read_decode
| data_type |    io_type    | cardinality | run_length | null_probability | data_size | Samples | CPU Time  | Noise | GPU Time  | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|-----------|---------------|-------------|------------|------------------|-----------|---------|-----------|-------|-----------|-------|------------------|-------------------|-------------------|
|    STRING | DEVICE_BUFFER |           0 |          1 |                0 | 536870912 |     19x | 26.434 ms | 0.33% | 26.430 ms | 0.33% |      20312875094 |         1.474 GiB |       482.854 MiB |
|    STRING | DEVICE_BUFFER |        1000 |          1 |                0 | 536870912 |    752x |  7.324 ms | 0.75% |  7.321 ms | 0.75% |      73337053811 |       553.638 MiB |        33.125 MiB |
|    STRING | DEVICE_BUFFER |           0 |         32 |                0 | 536870912 |     65x | 26.620 ms | 0.50% | 26.617 ms | 0.50% |      20170569924 |         1.474 GiB |       482.854 MiB |
|    STRING | DEVICE_BUFFER |        1000 |         32 |                0 | 536870912 |     67x |  7.555 ms | 0.48% |  7.552 ms | 0.48% |      71094241413 |       528.480 MiB |         3.678 MiB |
|    STRING | DEVICE_BUFFER |           0 |          1 |             0.01 | 536870912 |    304x | 27.627 ms | 0.58% | 27.623 ms | 0.58% |      19435511842 |         1.463 GiB |       478.839 MiB |
|    STRING | DEVICE_BUFFER |        1000 |          1 |             0.01 | 536870912 |    560x |  8.329 ms | 0.63% |  8.325 ms | 0.63% |      64489294562 |       570.596 MiB |        33.910 MiB |
|    STRING | DEVICE_BUFFER |           0 |         32 |             0.01 | 536870912 |    528x | 27.732 ms | 0.60% | 27.728 ms | 0.60% |      19361840196 |         1.463 GiB |       478.839 MiB |
|    STRING | DEVICE_BUFFER |        1000 |         32 |             0.01 | 536870912 |     62x |  8.127 ms | 0.33% |  8.123 ms | 0.32% |      66094341432 |       525.480 MiB |         3.696 MiB |

## parquet_read_chunks
|     T     |    io_type    | cardinality | run_length | chunk_read_limit | data_size | Samples |  CPU Time  | Noise |  GPU Time  | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|-----------|---------------|-------------|------------|------------------|-----------|---------|------------|-------|------------|-------|------------------|-------------------|-------------------|
|  INTEGRAL | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      5x | 158.052 ms | 0.23% | 158.048 ms | 0.23% |       3396881896 |       546.959 MiB |       477.515 MiB |
|  INTEGRAL | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      5x | 151.956 ms | 0.26% | 151.952 ms | 0.26% |       3533163434 |       219.219 MiB |       155.184 MiB |
|  INTEGRAL | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |      5x | 132.895 ms | 0.08% | 132.891 ms | 0.08% |       4039943681 |        68.402 MiB |        25.837 MiB |
|  INTEGRAL | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |      5x | 131.436 ms | 0.17% | 131.431 ms | 0.17% |       4084810668 |        64.924 MiB |        13.356 MiB |
|     FLOAT | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |     10x |  51.090 ms | 0.21% |  51.086 ms | 0.21% |      10509239645 |       525.910 MiB |       499.881 MiB |
|     FLOAT | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      8x |  70.539 ms | 0.26% |  70.535 ms | 0.26% |       7611422039 |       146.507 MiB |       107.339 MiB |
|     FLOAT | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |      9x |  59.252 ms | 0.18% |  59.248 ms | 0.19% |       9061409584 |        48.752 MiB |        22.186 MiB |
|     FLOAT | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |      9x |  58.833 ms | 0.34% |  58.829 ms | 0.34% |       9125979565 |        44.720 MiB |         9.064 MiB |
|     BOOL8 | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      5x | 819.254 ms | 0.03% | 819.248 ms | 0.03% |        655321799 |       248.081 MiB |        71.665 MiB |
|     BOOL8 | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      5x | 819.794 ms | 0.02% | 819.787 ms | 0.02% |        654890834 |       243.917 MiB |        70.101 MiB |
|     BOOL8 | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |      5x | 812.055 ms | 0.03% | 812.048 ms | 0.03% |        661132023 |       181.249 MiB |        21.573 MiB |
|     BOOL8 | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |      5x | 811.920 ms | 0.02% | 811.913 ms | 0.02% |        661241900 |       181.082 MiB |        21.467 MiB |
|   DECIMAL | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |     15x |  34.502 ms | 0.38% |  34.498 ms | 0.38% |      15562487057 |       701.863 MiB |       393.779 MiB |
|   DECIMAL | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |     12x |  42.544 ms | 0.44% |  42.540 ms | 0.44% |      12620429310 |       114.855 MiB |        70.082 MiB |
|   DECIMAL | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |     14x |  36.924 ms | 0.35% |  36.920 ms | 0.35% |      14541460474 |        48.992 MiB |        15.893 MiB |
|   DECIMAL | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |     15x |  34.085 ms | 0.36% |  34.081 ms | 0.36% |      15752616258 |        35.706 MiB |         5.885 MiB |
| TIMESTAMP | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      8x |  62.954 ms | 0.24% |  62.950 ms | 0.24% |       8528559570 |       704.025 MiB |       453.009 MiB |
| TIMESTAMP | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      9x |  56.144 ms | 0.31% |  56.139 ms | 0.31% |       9563203028 |       129.460 MiB |        90.129 MiB |
| TIMESTAMP | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |     11x |  49.239 ms | 0.17% |  49.235 ms | 0.17% |      10904343580 |        55.375 MiB |        19.251 MiB |
| TIMESTAMP | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |     11x |  46.428 ms | 0.23% |  46.424 ms | 0.23% |      11564466330 |        40.684 MiB |         7.786 MiB |
|  DURATION | DEVICE_BUFFER |           0 |          1 |           500000 | 536870912 |      9x |  56.393 ms | 0.21% |  56.389 ms | 0.21% |       9520846073 |       488.432 MiB |       349.613 MiB |
|  DURATION | DEVICE_BUFFER |        1000 |          1 |           500000 | 536870912 |      9x |  56.175 ms | 0.33% |  56.171 ms | 0.33% |       9557792609 |       127.570 MiB |        88.156 MiB |
|  DURATION | DEVICE_BUFFER |           0 |         32 |           500000 | 536870912 |     11x |  46.626 ms | 0.28% |  46.622 ms | 0.28% |      11515332805 |        42.128 MiB |        16.076 MiB |
|  DURATION | DEVICE_BUFFER |        1000 |         32 |           500000 | 536870912 |     11x |  46.482 ms | 0.27% |  46.478 ms | 0.27% |      11551179129 |        41.225 MiB |         7.565 MiB |
…